Refactored Activities to use Fragments
