+++ app/src/main/java/com/example/android/miwok/FamilyMembersActivity.java
@@ -1,163 +1,16 @@
 package com.example.android.miwok;
 
-import android.content.Context;
-import android.media.AudioManager;
-import android.media.MediaPlayer;
-import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
-
-import java.util.ArrayList;
+import android.os.Bundle;
 
 public class FamilyMembersActivity extends AppCompatActivity {
-
-    /** Handles playback of all sound files */
-    private MediaPlayer mMediaPlayer;
-
-    /** Handles audio focus when playing a sound file */
-    private AudioManager mAudioManager;
-
-    final AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener =
-            new AudioManager.OnAudioFocusChangeListener() {
-                public void onAudioFocusChange(int focusChange) {
-                    if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT) {
-                        // Pause playback because your Audio Focus was
-                        // temporarily stolen, but will be back soon.
-                        // i.e. for a phone call
-                        mMediaPlayer.pause();
-                    } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                        // Stop playback, because you lost the Audio Focus.
-                        // i.e. the user started some other playback app
-                        // Remember to unregister your controls/buttons here.
-                        // And release the kra — Audio Focus!
-                        // You’re done.
-                        releaseMediaPlayer();
-                        //mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-                    } else if (focusChange ==
-                            AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK) {
-                        // Lower the volume, because something else is also
-                        // playing audio over you.
-                        // i.e. for notifications or navigation directions
-                        // Depending on your audio playback, you may prefer to
-                        // pause playback here instead. You do you.
-                        mMediaPlayer.pause();
-                    } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                        // Resume playback, because you hold the Audio Focus
-                        // again!
-                        // i.e. the phone call ended or the nav directions
-                        // are finished
-                        // If you implement ducking and lower the volume, be
-                        // sure to return it to normal here, as well.
-                        mMediaPlayer.start();
-                    }
-                }
-            };
-
-    /** This listener gets triggered when the {@link MediaPlayer} has completed
-     * playing the audio file.
-     */
-    private MediaPlayer.OnCompletionListener mCompletionListener = new MediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(MediaPlayer mp) {
-            releaseMediaPlayer();
-        }
-    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.word_list);
-
-        // Create a list of words
-        final ArrayList<Word> words = new ArrayList<Word>();
-
-        words.add(new Word("father", "apa", R.drawable.family_father, R.raw.family_father));
-        words.add(new Word("mother", "ata", R.drawable.family_mother, R.raw.family_mother));
-        words.add(new Word("son", "angsi", R.drawable.family_son, R.raw.family_son));
-        words.add(new Word("daughter", "tune", R.drawable.family_daughter, R.raw.family_daughter));
-        words.add(new Word("older brother", "taachi", R.drawable.family_older_brother, R.raw.family_older_brother));
-        words.add(new Word("younger brother","chalitti", R.drawable.family_younger_brother, R.raw.family_younger_brother));
-        words.add(new Word("older sister", "tete", R.drawable.family_older_sister, R.raw.family_older_sister));
-        words.add(new Word("younger sister", "kolliti", R.drawable.family_younger_sister, R.raw.family_younger_sister));
-        words.add(new Word("grandmother", "ama", R.drawable.family_grandmother, R.raw.family_grandmother));
-        words.add(new Word("grandfather","paapa", R.drawable.family_grandfather, R.raw.family_grandfather));
-
-        WordAdapter adapter = new WordAdapter(this, words, R.color.category_family);
-
-        ListView listView = (ListView) findViewById(R.id.list);
-
-        listView.setAdapter(adapter);
-
-        //Set a click listener to play the audio when the list item is clicked on
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                //Get the {@link Word} object at the given position the user clicked on
-                Word word = words.get(position);
-                //Java will implicitly call the toString method on the object
-                Log.v("FamilyMembersActivity", "Current word: " + word);
-
-                // Release the media player if it currently exists because we are about to
-                // play a different sound file.
-                releaseMediaPlayer();
-
-                //Create and setup the {@link AudioManager} to request the audio focus
-                mAudioManager = (AudioManager) FamilyMembersActivity.this.getSystemService(Context.AUDIO_SERVICE);
-
-                //Request audio focus for playback
-                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
-                        // Use the music stream.
-                        AudioManager.STREAM_MUSIC,
-                        // Request audio focus for only a short time.
-                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
-
-
-                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-
-                    //Create and setup the {@link MediaPlayer} for the audio resource associated with the current word
-                    mMediaPlayer = MediaPlayer.create(FamilyMembersActivity.this, word.getAudioResourceId());
-
-                    // Start playback
-                    //Start the audio file
-                    mMediaPlayer.start();
-
-
-                    // Setup a listener on the media player, so that we can stop and release the
-                    //media player once the sounds has finished playing.
-                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
-                }
-
-            }
-        });
-
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        releaseMediaPlayer();
-    }
-
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer(){
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-        }
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.container, new FamilyMembersFragment())
+                .commit();
     }
 }
-
-
